Restições nas notifcações não imediatas corrigidas

+++ notificacoes/templatetags/notificacoes_extras.py
@@ -111,7 +111,20 @@ def atualizar_informacoes(user):
         info = InformacaoNotificacao.objects.filter(
                     recetor = utilizador_recetor)
         for x in info:
-            if timezone.now() >= x.data:
+            tmp = x.tipo
+            y = tmp.split()
+            type = y[0]
+            id = parseInt(y[1])
+            if type == "profile" or type == "register":
+                u = Utilizador.objects.get(id = id)
+                if u.valido == "False":
+                    pendente = True
+                elif u.valido == "True":
+                    pendente = False
+                else:
+                    x.delete()
+                    return ""
+            if timezone.now() >= x.data and pendente:
                 notify.send(sender=x.emissor, recipient=utilizador_recetor, verb=x.descricao, action_object=None,
                     target=None, level="info", description=x.titulo, public=True, timestamp=timezone.now())
                 x.delete()
